+++ src/app/organization/audit-trails/audit-trail.datasource.ts
@@ -5,7 +5,7 @@ import { CollectionViewer, DataSource } from '@angular/cdk/collections';
 import { Observable, BehaviorSubject } from 'rxjs';
 
 /** Custom Services */
-import { OrganizationService } from '../organization.service';
+import { AuditService } from './audit.service';
 
 /**
  * Audit Trails custom data source to implement server side filtering, pagination and sorting.
@@ -21,9 +21,9 @@ export class AuditTrailsDataSource implements DataSource<any> {
   public records$ = this.recordsSubject.asObservable();
 
   /**
-   * @param {OrganizationService} organizationService Organization Service.
+   * @param {AuditService} auditService Audit Service.
    */
-  constructor(private organizationService: OrganizationService) { }
+  constructor(private auditService: AuditService) { }
 
   /**
    * Gets audit trails on the basis of provided parameters and emits the value.
@@ -35,7 +35,7 @@ export class AuditTrailsDataSource implements DataSource<any> {
    */
   getAuditTrails(filterBy: any, orderBy: string = '', sortOrder: string = '', pageIndex: number = 0, limit: number = 10) {
     this.auditTrailsSubject.next([]);
-    this.organizationService.getAuditTrails(filterBy, orderBy, sortOrder, pageIndex * limit, limit)
+    this.auditService.getAuditTrails(filterBy, orderBy, sortOrder, pageIndex * limit, limit)
       .subscribe((auditTrails: any) => {
         this.recordsSubject.next(auditTrails.totalFilteredRecords);
         this.auditTrailsSubject.next(auditTrails.pageItems);